requirements.txt
aspose-cells-python==23.10.0
certifi==2023.7.22
charset-normalizer==3.3.1
et-xmlfile==1.1.0
greenlet==3.0.1
idna==3.4
lxml==4.9.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
xlrd==2.0.1
xmltodict==0.13.0
